requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
openai>=1.40.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
Jinja2>=3.1.4
numpy>=1.26.0
python-dotenv>=1.0.1
langid>=1.1.6
rapidfuzz>=3.9.6
duckduckgo_search>=6.1.7
trafilatura>=1.9.0
aiolimiter>=1.1.0
PyYAML>=6.0.1
typing_extensions>=4.9.0
langgraph>=0.2.35
langchain-core>=0.2.0
aiosqlite>=0.19.0
